insurance_rag_realestate/requirements.txt
fastapi
uvicorn
pinecone
google-genai
numpy
python-dotenv
